+++ src/main/java/ph/devcon/rapidpass/controllers/RegistryRestController.java
@@ -159,7 +159,7 @@ public class RegistryRestController {
     /**
      * Note that using the delete method will perform suspend access pass but without specifying the reason why the
      * rapid pass was suspended.
-     *
+     * <p>
      * To specify the reason why it was suspended, use the PUT request.
      */
     @DeleteMapping("/access-passes/{referenceId}")
@@ -210,7 +210,13 @@ public class RegistryRestController {
                 .body(responseBody);
     }
 
-    @GetMapping("/scanner-devices")
+    /**
+     * @param queryFilter
+     * @return
+     * @deprecated use {@link MobileDeviceController}
+     */
+    @Deprecated
+//    @GetMapping("/scanner-devices")
     public ResponseEntity<?> getScannerDevices(@RequestBody Optional<QueryFilter> queryFilter) {
 
         boolean isDisabled = true;
@@ -229,7 +235,13 @@ public class RegistryRestController {
         }
     }
 
-    @PostMapping("/scanner-devices")
+    /**
+     * @param deviceRequest
+     * @return
+     * @deprecated use {@link MobileDeviceController}
+     */
+    @Deprecated
+//    @PostMapping("/scanner-devices")
     public ResponseEntity<?> registerScannerDevice(@RequestBody MobileDevice deviceRequest) {
         ScannerDevice device = this.registryService.registerScannerDevice(deviceRequest);